Addressing PR comments.  Update variable names and index calls in population.py, moved unit test to proper directory
